+++ src/geodef/bayes.py
@@ -53,7 +53,12 @@ from geodef import backend, transforms
 from geodef.data import DataSet
 from geodef.fault import Fault
 from geodef.gradients import rect_greens, tri_greens
-from geodef.invert import _THETA_NAMES, LinearSystem, _projection_matrix
+from geodef.invert import (
+    _THETA_NAMES,
+    LinearSystem,
+    _projection_matrix,
+    _rank_positive_eigs,
+)
 
 _VALID_MODES = ("hierarchical", "weak", "profiled")
 _VALID_SLIP_MODES = ("hierarchical", "weak", "fixed")
@@ -729,7 +734,7 @@ class RectPosterior(_CollapsedPosterior):
                 self._lambda_fixed = float(smoothing_strength)
             else:
                 self._lambda_fixed = None
-            pos = eig[eig > 0]
+            pos = _rank_positive_eigs(eig)
             self._logdet_rank = len(pos)
             self._logdet_sum = float(np.sum(np.log(pos)))
 
@@ -1523,7 +1528,7 @@ class TriPosterior(_CollapsedPosterior):
                 self._lambda_fixed = float(smoothing_strength)
             else:
                 self._lambda_fixed = None
-            pos = eig[eig > 0]
+            pos = _rank_positive_eigs(eig)
             self._logdet_rank = len(pos)
             self._logdet_sum = float(np.sum(np.log(pos)))
         self._include_logdet = mode != "profiled"
@@ -1769,7 +1774,7 @@ class SlipPosterior:
         else:
             self._K = np.asarray(sys.LtL, dtype=np.float64)
             eig = np.abs(np.linalg.eigvalsh(self._K))
-            pos = eig[eig > 0]
+            pos = _rank_positive_eigs(eig)
             self._logdet_rank = len(pos)
             self._logdet_sum = float(np.sum(np.log(pos)))
             if mode == "fixed":